Restructure achievement commons
