+++ src/commons/achievement/control/GoalEditor.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+type GoalEditorProps = {};
+
+function GoalEditor(props: GoalEditorProps) {
+  return <div className="goal-editor"></div>;
+}
+
+export default GoalEditor;
