+++ src/commons/achievement/utils/AchievementLevel.tsx
@@ -2,7 +2,7 @@ import { ProgressBar } from '@blueprintjs/core';
 import React, { useState } from 'react';
 
 import { expPerLevel } from '../../../features/achievement/AchievementConstants';
-import AchievementMilestone from '../AchievementMilestone';
+import AchievementMilestone from './AchievementMilestone';
 
 type AchievementLevelProps = {
   studentExp: number;
