+++ src/styles/_achievementdashboard.scss
@@ -159,7 +159,7 @@
       }
     }
 
-    .list-container {
+    .task-container {
       @extend %container;
       height: 90%;
       list-style-type: none;
